Inception Submission 09/07/20
This folder contains both the presentation slides as well as the written deliverables. Since prof said that unless you don't have any major changes or point we can submit the slides for a written deliverable so we are submitting the slides only for written deliverable but we are have added some points in that slide for the written deliverable so we created a new slide with the name WrittenDeliverableDocument.